+++ services/runcommand/VERSION
@@ -1,1 +1,1 @@
-v1.9.1
+v1.10.0
